%zu printf specifier is not understood by MS Visual Studio 2008. Changing to %u.
git-svn-id: http://mongoose.googlecode.com/svn/trunk@223 a8d83a66-befe-11dd-b57b-7192f494c34b

+++ test/embed.c
@@ -49,7 +49,7 @@ test_get_var(struct mg_connection *conn, const struct mg_request_info *ri,
 	value = mg_get_var(conn, "my_var");
 	if (value != NULL) {
 		mg_printf(conn, "Value: [%s]\n", value);
-		mg_printf(conn, "Value size: [%zu]\n", strlen(value));
+		mg_printf(conn, "Value size: [%u]\n", (unsigned) strlen(value));
 		free(value);
 	}
 }
